added user authentication

+++ routes/api/cart.js
@@ -1,24 +1,32 @@
 const route = require('express').Router()
 const session = require('express-session')
-const passport = require('./passport')
+const passport = require('../../passport')
 const db = require('../../db')
 
 route.get('/', (req, res) => {
-    db.Cart.findAll({
-        include: [{
-            model: db.Product,
-            model: db.User
-        }]
-    })
-        .then((cart) => {
-            // console.log(cart)
-            res.status(200).send(cart)
+    if (req.user) {
+        db.Cart.findAll({
+            include: [{
+                model: db.Product,
+                model: db.User
+            }],
+            where: {
+                userId: req.body.id
+            }
         })
-        .catch((e) => {
-            res.status(500).send({
-                error: "Could not retrieve Cart Product(s)" + e
+            .then((cart) => {
+                // console.log(cart)
+                res.status(200).send(cart)
             })
-        })
+            .catch((e) => {
+                res.status(500).send({
+                    error: "Could not retrieve Cart Product(s)" + e
+                })
+            })
+    }
+    else {
+        res.send("Please Login First!")
+    }
 })
 
 route.post('/', (req, res) => {
@@ -27,27 +35,34 @@ route.post('/', (req, res) => {
     //         error: "Input is not a valid number"
     //     })
     // }
-    db.Product.find({
-        where: {
-            id: req.body.productId
-        }
-    }).then((product) => {
-        db.Cart.create({
-            productId: product.id,
-            quantity: 1
-        })
-    })
-        .then((cart) => {
-            res.status(201).send(cart)
-        })
-        .catch((e) => {
-            res.status(501).send({
-                error: "Could not add new cart object"
+    if (req.user) {
+        db.Product.find({
+            where: {
+                id: req.body.productId
+            }
+        }).then((product) => {
+            db.Cart.create({
+                productId: product.id,
+                quantity: 1,
+                userId: req.user.id
             })
         })
+            .then((cart) => {
+                res.status(201).send(cart)
+            })
+            .catch((e) => {
+                res.status(501).send({
+                    error: "Could not add new cart object"
+                })
+            })
+    }
+    else {
+        res.redirect('login.html')
+    }
 })
 
 route.delete('/:id', (req, res) => {
+    if(req.user){
     db.Cart.destroy({
         where: {
             id: req.params.id
@@ -58,41 +73,47 @@ route.delete('/:id', (req, res) => {
                 error: ("Could not delete cart item")
             })
         })
+    }
 })
 
 route.put('/', (req, res) => {
-    if (req.body.sign == 1) {
-        db.Cart.increment('quantity', {
-            where: {
-                id: req.body.id
-            }
-        })
-            .catch((e) => {
-                res.status(500).send({ error: "Could not update Cart Product" })
-            })
-    }
-    if (req.body.sign == -1) {
-        db.Cart.findById(req.body.id)
-            .then((cart) => {
-                if (cart.quantity == 1) {
-                    db.Cart.destroy({
-                        where: {
-                            id: req.body.id
-                        }
-                    })
+    if (req.user) {
+        if (req.body.sign == 1) {
+            db.Cart.increment('quantity', {
+                where: {
+                    id: req.body.id
                 }
-                else {
-                    db.Cart.decrement('quantity', {
-                        where: {
-                            id: req.body.id
-                        }
-                    })
-                }
-            }).catch((e) => {
-                res.status(500).send({ error: "Could not update Cart Product" })
             })
+                .catch((e) => {
+                    res.status(500).send({ error: "Could not update Cart Product" })
+                })
+        }
+        if (req.body.sign == -1) {
+            db.Cart.findById(req.body.id)
+                .then((cart) => {
+                    if (cart.quantity == 1) {
+                        db.Cart.destroy({
+                            where: {
+                                id: req.body.id
+                            }
+                        })
+                    }
+                    else {
+                        db.Cart.decrement('quantity', {
+                            where: {
+                                id: req.body.id
+                            }
+                        })
+                    }
+                }).catch((e) => {
+                    res.status(500).send({ error: "Could not update Cart Product" })
+                })
+        }
+        res.status(500).send({ error: "Bad request" })
     }
-    res.status(500).send({ error: "Bad request" })
+    else {
+        res.send("Please Login First!")
+    }
 })
 
 exports = module.exports = route
